requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
django-cors-headers
bcrypt
argon2-cffi